refactor: remove parent parser mod
