--- src/parsers/mod.rs
@@ -1,3 +1,0 @@
-pub mod error;
-pub mod generators;
-pub mod tokenizer;
